// src/app/shared/model/NegativeCustChangeTrxObj.Model.ts
export class NegativeCustChangeTrxObj {
    NegativeCustChangeTrxId: any;
    NegativeCustId: any;
    TrxNo: any;
    MrTrxStatCode: any;
    MrNegCustTypeCode: any;
    MrNegCustSourceCode: any;
    NegCustCause: any;
    Notes: any;
    RfaNo: any;
    ReqDt: any;
    ApvDt: any;
    ExeDt: any;
    RowVersion: any;
    constructor(){ this.NegativeCustChangeTrxId = 0; this.RowVersion = ''; }
}